+++ PXCTestKitTests/Fixtures/Crash.source/Crash/AppDelegate.h
@@ -1,0 +1,15 @@
+//
+//  AppDelegate.h
+//  Crash
+//
+//  Created by Johannes Plunien on 3/12/2016.
+//  Copyright © 2016 Johannes Plunien. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface AppDelegate : UIResponder <UIApplicationDelegate>
+
+@property (strong, nonatomic) UIWindow *window;
+
+@end