+++ PXCTestKitTests/Fixtures/Crash.source/Crash/AppDelegate.m
@@ -1,0 +1,17 @@
+//
+//  AppDelegate.m
+//  Crash
+//
+//  Created by Johannes Plunien on 3/12/2016.
+//  Copyright © 2016 Johannes Plunien. All rights reserved.
+//
+
+#import "AppDelegate.h"
+
+@interface AppDelegate ()
+
+@end
+
+@implementation AppDelegate
+
+@end