+++ PXCTestKitTests/Fixtures/Crash.source/Crash/Thing.h
@@ -1,0 +1,15 @@
+//
+//  Thing.h
+//  Crash
+//
+//  Created by Johannes Plunien on 3/12/2016.
+//  Copyright © 2016 Johannes Plunien. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+@interface Thing : NSObject
+
++ (void)crash;
+
+@end