+++ PXCTestKitTests/Fixtures/Crash.source/Crash/main.m
@@ -1,0 +1,16 @@
+//
+//  main.m
+//  Crash
+//
+//  Created by Johannes Plunien on 3/12/2016.
+//  Copyright © 2016 Johannes Plunien. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+#import "AppDelegate.h"
+
+int main(int argc, char * argv[]) {
+    @autoreleasepool {
+        return UIApplicationMain(argc, argv, nil, NSStringFromClass([AppDelegate class]));
+    }
+}